refactor: StoreStatus moved to http pacakge
- Node and StoreStatus moved to http package
- Json marshal is handled by http service
- store package import http package for StoreStatus and Node
